ref: build ExpandedKey from lnd signature message
Changes the generation of the ExpandedKey that is used
to all ldk offers signature operations. With this changes
we don't require any other signature operation from lnd side.
Building offer signatures this way maintains statelessness of
lndk, makes it restart-proof and in any case of leaking it does
not makes any problems on funds.