handler: respond to invoice requests with invoices
Whenver an offer message arrives we verify that is from a offer
generated from the lndk keys. Then we create an invoice on LND
side to finally respond the newly generated Invoice through
an oniont message.
For given invoice we add a Nonce and Hmac to authenticate in case
of any error.
Finally, we add a simple integration test of generating an offer
andh then paying it.